requirements.txt
Pillow==9.1.1
pytest==7.1.2